refactor: STORIF-335 - CreateBucketDrawer replaced by OMC_CreateBucketDrawer
